LPD-75112 Rename site-config-export to site-settings-export and remove redundant properties
