+++ modules/test/playwright/tests/configuration-admin-web/system-settings-with-ui/systemSettingsWithUI.spec.ts
