+++ modules/test/playwright/tests/configuration-admin-web/system-settings-with-ui/test.properties
