+++ test.properties
@@ -4836,8 +4836,8 @@
         configuration-admin-web.main,\
         configuration-admin-web.override-properties-only,\
         configuration-admin-web.override-with-osgi,\
-        configuration-admin-web.site-config-export,\
-        configuration-admin-web.system-settings-via-ui,\
+        configuration-admin-web.site-settings-export,\
+        configuration-admin-web.system-settings-with-ui,\
         feature-flag-web.main,\
         frontend-css-cadmin-web.main,\
         frontend-js-item-selector-web.main,\
